[fix] 리스트 내부 onClick 이벤트 발생안하는 문제 해결 #40 - List를 따로 컴포넌트로 분리하였음
